requirements.txt
aerich==0.9.0 ; python_version >= "3.9" and python_version < "4.0"
aiohappyeyeballs==2.6.1 ; python_version >= "3.9" and python_version < "4.0"
aiohttp-retry==2.9.1 ; python_version >= "3.9" and python_version < "4.0"
aiohttp==3.11.18 ; python_version >= "3.9" and python_version < "4.0"
aiosignal==1.3.2 ; python_version >= "3.9" and python_version < "4.0"
aiosqlite==0.21.0 ; python_version >= "3.9" and python_version < "4.0"
amqp==5.3.1 ; python_version >= "3.9" and python_version < "4.0"
annotated-types==0.7.0 ; python_version >= "3.9" and python_version < "4.0"
anyio==4.9.0 ; python_version >= "3.9" and python_version < "4.0"
async-timeout==5.0.1 ; python_version >= "3.9" and python_full_version < "3.11.3"
asyncclick==8.1.8 ; python_version >= "3.9" and python_version < "4.0"
asyncmy==0.2.10 ; python_version >= "3.9" and python_version < "4.0"
asyncpg==0.30.0 ; python_version >= "3.9" and python_version < "4.0"
attrs==25.3.0 ; python_version >= "3.9" and python_version < "4.0"
billiard==4.2.1 ; python_version >= "3.9" and python_version < "4.0"
celery==5.5.2 ; python_version >= "3.9" and python_version < "4.0"
certifi==2025.4.26 ; python_version >= "3.9" and python_version < "4.0"
charset-normalizer==3.4.2 ; python_version >= "3.9" and python_version < "4.0"
click-didyoumean==0.3.1 ; python_version >= "3.9" and python_version < "4.0"
click-plugins==1.1.1 ; python_version >= "3.9" and python_version < "4.0"
click-repl==0.3.0 ; python_version >= "3.9" and python_version < "4.0"
click==8.1.8 ; python_version >= "3.9" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.9" and python_version < "4.0" and platform_system == "Windows"
dictdiffer==0.9.0 ; python_version >= "3.9" and python_version < "4.0"
distro==1.9.0 ; python_version >= "3.9" and python_version < "4.0"
eval-type-backport==0.2.2 ; python_version == "3.9"
exceptiongroup==1.3.0 ; python_version >= "3.9" and python_version < "3.11"
fastapi==0.115.12 ; python_version >= "3.9" and python_version < "4.0"
frozenlist==1.6.0 ; python_version >= "3.9" and python_version < "4.0"
h11==0.16.0 ; python_version >= "3.9" and python_version < "4.0"
httpcore==1.0.9 ; python_version >= "3.9" and python_version < "4.0"
httpx==0.28.1 ; python_version >= "3.9" and python_version < "4.0"
idna==3.10 ; python_version >= "3.9" and python_version < "4.0"
iso8601==2.1.0 ; python_version >= "3.9" and python_version < "4.0"
jiter==0.10.0 ; python_version >= "3.9" and python_version < "4.0"
kombu==5.5.3 ; python_version >= "3.9" and python_version < "4.0"
multidict==6.4.3 ; python_version >= "3.9" and python_version < "4.0"
openai==1.82.0 ; python_version >= "3.9" and python_version < "4.0"
prompt-toolkit==3.0.51 ; python_version >= "3.9" and python_version < "4.0"
propcache==0.3.1 ; python_version >= "3.9" and python_version < "4.0"
pydantic-core==2.33.2 ; python_version >= "3.9" and python_version < "4.0"
pydantic==2.11.4 ; python_version >= "3.9" and python_version < "4.0"
pyjwt==2.10.1 ; python_version >= "3.9" and python_version < "4.0"
pypika-tortoise==0.5.0 ; python_version >= "3.9" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.9" and python_version < "4.0"
python-dotenv==1.1.0 ; python_version >= "3.9" and python_version < "4.0"
python-multipart==0.0.20 ; python_version >= "3.9" and python_version < "4.0"
pytz==2025.2 ; python_version >= "3.9" and python_version < "4.0"
redis==6.1.0 ; python_version >= "3.9" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.9" and python_version < "4.0"
six==1.17.0 ; python_version >= "3.9" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.9" and python_version < "4.0"
starlette==0.46.2 ; python_version >= "3.9" and python_version < "4.0"
tomli==2.2.1 ; python_version >= "3.9" and python_version < "4.0"
tomlkit==0.13.2 ; python_version >= "3.9" and python_version < "4.0"
tortoise-orm==0.25.0 ; python_version >= "3.9" and python_version < "4.0"
tqdm==4.67.1 ; python_version >= "3.9" and python_version < "4.0"
twilio==9.6.1 ; python_version >= "3.9" and python_version < "4.0"
typing-extensions==4.13.2 ; python_version >= "3.9" and python_version < "4.0"
typing-inspection==0.4.0 ; python_version >= "3.9" and python_version < "4.0"
tzdata==2025.2 ; python_version >= "3.9" and python_version < "4.0"
urllib3==2.4.0 ; python_version >= "3.9" and python_version < "4.0"
uvicorn==0.34.2 ; python_version >= "3.9" and python_version < "4.0"
vine==5.1.0 ; python_version >= "3.9" and python_version < "4.0"
wcwidth==0.2.13 ; python_version >= "3.9" and python_version < "4.0"
yarl==1.20.0 ; python_version >= "3.9" and python_version < "4.0"
